Fix EOEPCA-939
Fix the Deploy/Undeploy processes when no metadb is defined

+++ zoo-project/zoo-services/utils/open-api/dru/UndeployProcess.py
@@ -88,25 +88,29 @@ class UndeployService(object):
 
 
     def remove_service(self):
-        import psycopg2
-        import psycopg2.extensions
-        psycopg2.extensions.register_type(psycopg2.extensions.UNICODE)
-        conn = psycopg2.connect("host=%s port=%s dbname=%s user=%s password=%s" % (self.conf["metadb"]["host"], self.conf["metadb"]["port"], self.conf["metadb"]["dbname"], self.conf["metadb"]["user"], self.conf["metadb"]["password"]))
-        cur = conn.cursor()
-        if "auth_env" in self.conf:
-            self.user=self.conf["auth_env"]["user"]
-        else:
-            self.user="anonymous"
-        cur.execute("DELETE FROM collectiondb.ows_process WHERE identifier='%s' AND user_id=(select id from public.users where name=$q$%s$q$)" % (self.get_process_identifier(),self.user))
-        conn.commit()
-        conn.close()
+
+        if "metadb" in self.conf:
+            import psycopg2
+            import psycopg2.extensions
+            psycopg2.extensions.register_type(psycopg2.extensions.UNICODE)
+            conn = psycopg2.connect("host=%s port=%s dbname=%s user=%s password=%s" % (self.conf["metadb"]["host"], self.conf["metadb"]["port"], self.conf["metadb"]["dbname"], self.conf["metadb"]["user"], self.conf["metadb"]["password"]))
+            cur = conn.cursor()
+            if "auth_env" in self.conf:
+                self.user=self.conf["auth_env"]["user"]
+            else:
+                self.user="anonymous"
+            cur.execute("DELETE FROM collectiondb.ows_process WHERE identifier='%s' AND user_id=(select id from public.users where name=$q$%s$q$)" % (self.get_process_identifier(),self.user))
+            conn.commit()
+            conn.close()
+
         service_folder = os.path.join(self.zooservices_folder, self.service_identifier)
         if os.path.isdir(service_folder):
             shutil.rmtree(service_folder)
 
-        service_configuration_file = f"{service_folder}.zcfg"
-        if os.path.exists(service_configuration_file):
-            os.remove(service_configuration_file)
+        for i in [".zcfg",".json"]:
+            service_configuration_file = f"{service_folder}"+i
+            if os.path.exists(service_configuration_file):
+                os.remove(service_configuration_file)
